Rename index utils to transactions
